requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
email-validator==2.1.0
python-dateutil==2.8.2
flask-limiter==3.3.3